+++ docs/CHANGELOG.md
@@ -21,7 +21,7 @@ Alle viktige endringer i NISSY-scriptene vil bli dokumentert i denne filen.
   - Ny modus: Hvis 2+ ventende bestillinger er merket, anses de som at de skal samkjøres, og søker samlet etter ressurser på pågående oppdrag.
   - Kan velge hvilke man ønsker å samkjøre og merke disse for videre justering / tildeling.
 - 🔔 **Overvåk-Ventende** 
-  - Lagt til i basic- og advanced-loader.
+  - Lagt til i basic- og advanced-loader. Installeres som en knapp over ventende oppdrag.
 
 ## 🚀 [3.5.1] - 10.02.2026